audio: note durations (do2, do/4) + one play at a time
Durations (LANGUAGE.md §13, decided #43): a note carries an optional
suffix attached with no space — a whole number = beats (do2 = 2), /N =
one-N-th of a beat (do/4 = 0.25). The lexer reads do2 / do/4 as a single
note token (so do2 isn't a variable name), but a lookalike like 'dog'
(base isn't a note) stays a name and x/2 stays division. beats goes u16 →
f32 through Toon/Voice/the FFI JSON and the Swift synth. 3 new core tests.

One play at a time (#52 follow-up): the synth schedules with .interrupts,
so a new play cancels the in-flight buffer instead of queueing — 'play' in
a loop no longer stacks a minutes-long backlog.

Verified end-to-end on the sim: do2/mi/2/sol/4 highlight as single gold
note tokens, an 8x play-in-loop + final play runs crash-free. 33 core +
3 ffi tests pass.

+++ MaakSwift/Maak/Maak/Schildpad.swift
@@ -21,7 +21,7 @@ struct RawEvent: Decodable {
 
 struct RawVoice: Decodable {
     var hz: Float?      // null = a rest (stilte)
-    var beats: Int
+    var beats: Double   // fractional beats (do2 = 2, do/4 = 0.25)
     var osc: String
     var env: String
 }
